fix(agent): the cpu check was working with a 32bits counter
That's not enough, on  a surver this counter is a uint64 and when
the check had to read a such counter, it failed.

+++ agent/native_linux/src/check_cpu.cc
@@ -48,7 +48,7 @@ per_cpu_time::per_cpu_time(const std::string_view& line) {
   auto to_fill = _metrics.begin();
   auto end = _metrics.end();
   for (++field_iter; field_iter != split_res.end(); ++field_iter, ++to_fill) {
-    unsigned counter;
+    uint64_t counter;
     if (!absl::SimpleAtoi(*field_iter, &counter)) {
       throw std::invalid_argument("not a number");
     }
